Move test for tempclient
